ML-446 Segregated the distance and correlation metrics
