+++ requirements.txt
@@ -1,10 +1,9 @@
+dataclasses==0.6;python_version=='3.6'
+dcor
 flake8
 mypy
 numpy
 pandas
 pytest
 sklearn
-pytest
-dcor
 types-dataclasses; python_version < '3.7'
-dataclasses==0.6;python_version=='3.6'
