+++ setup.cfg
@@ -16,12 +16,12 @@ project_urls =
 packages = find:
 zip_safe = False
 install_requires =
+    dataclasses ~= 0.6; python_version <'3.7'
+    dcor >= 0.5.3
     numpy >= 1.18.4, < 1.20.0
     pandas ~= 1.1
     scikit_learn ~= 0.23
     scipy ~= 1.5
-    dcor >= 0.5.3
-    dataclasses ~= 0.6; python_version <'3.7'
     types-dataclasses; python_version < '3.7'
 python_requires = >=3.6
