+++ src/synthesized_insight/metrics/__init__.py
@@ -1,14 +1,16 @@
-from .base import TwoColumnMetric, TwoDataFrameMetric
-from .metrics import (BinomialDistance, CramersV, DistanceCNCorrelation,
-                      DistanceNNCorrelation, EarthMoversDistance,
-                      EarthMoversDistanceBinned, HellingerDistance,
-                      JensenShannonDivergence, KendallTauCorrelation,
-                      KolmogorovSmirnovDistance, KruskalWallis,
-                      KullbackLeiblerDivergence, Mean, MetricStatistics, Norm,
-                      R2Mcfadden, SpearmanRhoCorrelation, StandardDeviation)
+from .base import (BinnedMetricStatistics, MetricStatistics, TwoColumnMetric,
+                   TwoDataFrameMetric)
+from .correlation import (CramersV, DistanceCNCorrelation,
+                          DistanceNNCorrelation, KendallTauCorrelation,
+                          R2Mcfadden, SpearmanRhoCorrelation)
+from .distance import (BinomialDistance, EarthMoversDistance,
+                       EarthMoversDistanceBinned, HellingerDistance,
+                       JensenShannonDivergence, KolmogorovSmirnovDistance,
+                       KruskalWallis, KullbackLeiblerDivergence, Mean, Norm,
+                       StandardDeviation)
 
 __all__ = ['TwoColumnMetric', 'TwoDataFrameMetric', 'CramersV', 'EarthMoversDistance', 'KendallTauCorrelation',
            'KolmogorovSmirnovDistance', 'Mean', 'SpearmanRhoCorrelation', 'StandardDeviation',
            'R2Mcfadden', 'KruskalWallis', 'DistanceNNCorrelation', 'DistanceCNCorrelation',
            'KullbackLeiblerDivergence', 'HellingerDistance', 'JensenShannonDivergence', 'Norm',
-           "BinomialDistance", "MetricStatistics", "EarthMoversDistanceBinned"]
+           'BinomialDistance', 'MetricStatistics', 'BinnedMetricStatistics', 'EarthMoversDistanceBinned']
